# public/openapi/agent-commerce-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Veyra Agent API",
    "version": "1.5.0",
    "description": "Machine API for verified Veyra workflows, Project 360 free discovery and explicit quoting, continuous trust monitoring, alerts, signed webhooks, and public trust badges."
  },
  "servers": [
    {
      "url": "https://agent-commerce-six.vercel.app",
      "description": "Arc Testnet Production Server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/agent/v1/project-360/discoveries/{discoveryId}/quote": {
      "post": {
        "summary": "Create an immutable Project 360 quote",
        "description": "Binds explicit candidate IDs and modules to the current discovery revision. The response lists every module price, total price, expected coverage, and incomplete-data warnings. No unselected candidate is included. Requires quotes:create scope and Idempotency-Key.",
        "operationId": "createProject360Quote",
        "parameters": [
          { "$ref": "#/components/parameters/IdempotencyKey" },
          {
            "name": "discoveryId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "pattern": "^dsc_[0-9a-f]{20}$" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project360QuoteRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Idempotent quote replay",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project360QuoteResponse" } } }
          },
          "201": {
            "description": "Immutable quote created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project360QuoteResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "aac_live_...",
        "description": "Veyra Agent API Bearer credential token (`aac_live_...`)"
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Unique key to ensure request idempotency (e.g. UUIDv4)",
        "schema": {
          "type": "string",
          "example": "idemp_9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
        }
      },
...
    },
    "schemas": {
...
      "MachineError": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "retryable", "requestId"],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "agent_trust_input_required",
                  "agent_not_found",
                  "agent_access_denied",
                  "agent_registry_unavailable",
                  "agent_trust_service_unavailable",
                  "contract_not_found",
                  "contract_provider_unavailable",
                  "endpoint_invalid",
                  "endpoint_private_network_blocked",
                  "endpoint_unreachable",
                  "endpoint_response_too_large",
                  "insufficient_trust_evidence",
                  "invalid_wallet",
                  "invalid_repository",
                  "repository_not_found",
                  "repository_inaccessible",
                  "credential_missing",
                  "credential_revoked",
                  "scope_denied",
                  "workflow_disabled",
                  "quote_expired",
                  "quote_not_found",
                  "quote_already_used",
                  "idempotency_key_missing",
                  "idempotency_conflict",
                  "idempotency_in_progress",
                  "idempotency_store_unavailable",
                  "invalid_request",
                  "payment_required",
                  "payment_invalid",
                  "spending_limit_exceeded",
                  "watchlist_invalid",
                  "watchlist_not_found",
                  "watchlist_limit_exceeded",
                  "recheck_not_found",
                  "recheck_conflict",
                  "recheck_in_progress",
                  "monitoring_unavailable",
                  "run_not_found",
                  "report_not_found",
                  "report_not_ready",
                  "report_generation_failed",
                  "verification_pending",
                  "provider_unavailable",
                  "rate_limited",
                  "internal_error",
                  "api_quality_service_not_found",
                  "api_quality_service_unavailable",
                  "api_quality_insufficient_services",
                  "api_quality_too_many_services",
                  "api_quality_no_observations",
                  "api_quality_observation_store_unavailable",
                  "api_quality_comparison_incompatible",
                  "api_quality_probe_budget_exceeded"
                ],
                "example": "api_quality_service_not_found",
                "description": "Stable Machine API error code."
              },
              "message": {
                "type": "string",
                "example": "Enter a valid GitHub repository in owner/repository format."
              },
              "retryable": {
                "type": "boolean",
                "example": false
              },
              "requestId": {
                "type": "string",
                "example": "req_8f12a45b7e90"
              }
            }
          }
        }
      },
...
      "Project360Module": {
        "type": "string",
        "enum": [
          "github_due_diligence",
          "agent_trust_report",
          "treasury_health",
          "paid_api_quality",
          "arc_contract_analysis"
        ]
      },
...
      "Project360QuoteRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["revision", "selectedCandidateIds", "modules"],
        "properties": {
          "revision": { "type": "integer", "minimum": 1 },
          "selectedCandidateIds": {
            "type": "array",
            "minItems": 1,
            "maxItems": 5,
            "uniqueItems": true,
            "items": { "type": "string", "pattern": "^src_[0-9a-f]{20}$" }
          },
          "modules": {
            "type": "array",
            "minItems": 1,
            "maxItems": 5,
            "uniqueItems": true,
            "items": { "$ref": "#/components/schemas/Project360Module" }
          }
        }
      },
      "Project360QuoteSelection": {
        "type": "object",
        "required": ["discoveryId", "discoveryRevision", "discoverySnapshotHash", "selectionHash", "expectedCoverage", "expectedCoverageLabel", "warnings", "confirmedSources", "selectedModules", "lineItems", "pricing", "canonicalInput"],
        "properties": {
          "discoveryId": { "type": "string", "pattern": "^dsc_[0-9a-f]{20}$" },
          "discoveryRevision": { "type": "integer" },
          "discoverySnapshotHash": { "type": "string" },
          "selectionHash": { "type": "string" },
          "expectedCoverage": {
            "type": "object",
            "required": ["selected", "total"],
            "properties": {
              "selected": { "type": "integer", "minimum": 1, "maximum": 5 },
              "total": { "type": "integer", "enum": [5] }
            }
          },
          "expectedCoverageLabel": { "type": "string", "example": "3 of 5 modules" },
          "warnings": { "type": "array", "items": { "type": "string" } },
          "confirmedSources": { "type": "array", "items": { "type": "object" } },
          "selectedModules": { "type": "array", "items": { "$ref": "#/components/schemas/Project360Module" } },
          "lineItems": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["module", "label", "serviceSlugs", "priceUsdc", "sharedEvidence"],
              "properties": {
                "module": { "type": "string" },
                "label": { "type": "string" },
                "serviceSlugs": { "type": "array", "items": { "type": "string" } },
                "priceUsdc": { "type": "number", "minimum": 0 },
                "sharedEvidence": { "type": "boolean" }
              }
            }
          },
          "pricing": {
            "type": "object",
            "required": ["moduleSubtotalUsdc", "platformFeeUsdc", "totalUsdc", "amountDueUsdc"],
            "properties": {
              "moduleSubtotalUsdc": { "type": "number", "minimum": 0 },
              "platformFeeUsdc": { "type": "number", "minimum": 0 },
              "totalUsdc": { "type": "number", "minimum": 0 },
              "amountDueUsdc": { "type": "number", "minimum": 0 }
            }
          },
          "canonicalInput": { "type": "string" }
        }
      },
      "Project360QuoteResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/WorkflowQuoteResponse" },
          {
            "type": "object",
            "required": ["project360"],
            "properties": { "project360": { "$ref": "#/components/schemas/Project360QuoteSelection" } }
          }
        ]
      },
...
      "WorkflowQuoteResponse": {
        "type": "object",
        "required": ["quoteId", "workflow", "totalUsdc", "sponsored", "expiresAt", "requiredPayment"],
        "properties": {
          "quoteId": {
            "type": "string",
            "example": "qte_9f81a2b3c4d5"
          },
          "workflow": {
            "type": "string",
            "example": "github_due_diligence"
          },
          "repository": {
            "type": "object",
            "nullable": true,
            "properties": {
              "fullName": { "type": "string", "example": "circlefin/agent-commerce" },
              "canonicalUrl": { "type": "string", "example": "https://github.com/circlefin/agent-commerce" }
            }
          },
          "totalUsdc": {
            "type": "number",
            "example": 0.002
          },
          "sponsored": {
            "type": "boolean",
            "example": true
          },
          "checkout": {
            "type": "object",
            "properties": {
              "mode": { "type": "string", "enum": ["sponsored", "arc_transaction"], "example": "sponsored" },
              "asset": { "type": "string", "example": "USDC" },
              "network": { "type": "string", "example": "arc-testnet" }
            }
          },
          "downstreamSettlement": {
            "type": "string",
            "example": "server_side_x402"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-24T22:15:00.000Z"
          },
          "requiredPayment": {
            "type": "object",
            "properties": {
              "network": { "type": "string", "example": "arc-testnet" },
              "asset": { "type": "string", "example": "USDC" },
              "amount": { "type": "number", "example": 0.0 },
              "treasuryAddress": { "type": "string", "example": "0x742d35Cc6634C0532925a3b844Bc454e4438f44e" },
              "chainId": { "type": "number", "example": 5042002 }
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid input or repository reference",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MachineError"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid Bearer token credential",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MachineError"
            }
          }
        }
      },
      "PaymentRequired": {
        "description": "Payment authorization required for paid workflow quote",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MachineError"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Credential scope denied or workflow disabled",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MachineError"
            }
          }
        }
      },
      "NotFound": {
        "description": "Quote, run, or report not found or unowned",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MachineError"
            }
          }
        }
      },
      "Conflict": {
        "description": "Idempotency conflict, mutation already in progress, or quote already executed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MachineError"
            }
          }
        }
      },
...
      "ServiceUnavailable": {
        "description": "Downstream x402 service or provider unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MachineError"
            }
          }
        }
      }
    }
  }
}